refactor: update UserDao, JdbcTemplate applied DbcTemplate
